refactor(script): split interpreter dispatch into operation modules (#44)
Move per-opcode implementations into category files under
Sources/BSV/Script/Interpreter/Operations/: DataPush, StackOps,
Arithmetic, Bitwise, Crypto, FlowControl, Splice. Interpreter.swift
retains the execution loop and dispatch table only.

Closes #44
